+++ compiler/include/ConstantLoad.h
@@ -32,12 +32,12 @@ class ConstantLoadInsn : public NonTerminatorInsn {
   private:
     TypeTag typeTag;
     LLVMValueRef getNewString(LLVMModuleRef &modRef);
-    std::variant<int, double, bool, std::string> value;
+    std::variant<uint64_t, double, bool, std::string> value;
     std::unique_ptr<llvm::GlobalVariable> globalStringValue;
 
   public:
     ConstantLoadInsn() = delete;
-    ConstantLoadInsn(const Operand &lhs, std::shared_ptr<BasicBlock> currentBB, int intVal);
+    ConstantLoadInsn(const Operand &lhs, std::shared_ptr<BasicBlock> currentBB, uint64_t intVal);
     ConstantLoadInsn(const Operand &lhs, std::shared_ptr<BasicBlock> currentBB, float floatVal);
     ConstantLoadInsn(const Operand &lhs, std::shared_ptr<BasicBlock> currentBB, bool boolVal);
     ConstantLoadInsn(const Operand &lhs, std::shared_ptr<BasicBlock> currentBB, std::string str);
